requirements.txt
qrcode
Pillow   #gerar imagens do Qrcode
python-dotenv   #carregar variáveis de ambiente
os
